+++ R/chrono.subsets.R
@@ -192,7 +192,6 @@ chrono.subsets <- function(data, tree, method, time, model, inc.nodes = FALSE, F
     all_methods <- c("discrete", "d", "continuous", "c")
     ## method must be a character string
     check.class(method, "character")
-    method <- tolower(method)
     ## method must have only one element
     check.length(method, 1, paste(" argument must be one of the following: ", paste(all_methods, collapse = ", "), ".", sep = ""))
     ## method must be either "discrete", "d", "continuous", or "c"